+++ slsDetectorSoftware/multiSlsDetector/multiSlsDetector.h
@@ -1440,7 +1440,7 @@ private:
 	 * @param filename address of file name
 	 */
 	int getData(const int isocket, const bool masking, int* image, const int size,
-			uint64_t &acqIndex, uint64_t &frameIndex, uint64_t &subframeIndex, string &filename);
+			uint64_t &acqIndex, uint64_t &frameIndex, uint32_t &subframeIndex, string &filename);
 
 	/** Ensures if sockets created successfully */
 	bool dataSocketsStarted;
